requirements.txt
django==5.2.4
djangorestframework==3.16.1
djangorestframework-simplejwt==5.2.1
django-cors-headers==4.7.0
drf-yasg==1.21.10
psycopg2-binary==2.9.10
python-dotenv==1.0.0
pytest==7.4.0
pytest-django==4.5.2
coverage==7.2.2
